refactor(fw): move hmac_hash to the shared ddi layer
Relocate the `HsmVaultKeyKind -> HsmHashAlgo` HMAC mapping out of the
MBOR-only `mbor::from_pal` module up to the codec-neutral `ddi` module
(alongside `recover_bk_boot`), so the upcoming TBOR HMAC commands can reuse
it without depending on `mbor`.  Pure refactor: MBOR's `Hmac` handler now
calls `crate::ddi::hmac_hash`; behavior is unchanged.

Validation: MBOR HMAC emu 36/36 (fixed + var-len HMAC, all SHA variants,
sign-permission, KBKDF-derived, masked-key HMAC); fmt + clippy clean.

Copilot-Session: 0b09e50a-a9be-4bae-b347-d42dc775a258
